# Performance: batched CLIP embeddings, cached model, parallel + early-termination re-rank