--- a/image_recommender/pipeline/search_pipeline.py
+++ b/image_recommender/pipeline/search_pipeline.py
@@ -6,6 +6,7 @@ from PIL import Image
 import numpy as np
 from concurrent.futures import ThreadPoolExecutor, as_completed
 import multiprocessing
+from heapq import heappush, heappushpop, nlargest
 
 from image_recommender.data.loader import load_image, preprocess_image
 from image_recommender.similarity.similarity_embedding import compute_clip_embedding, load_annoy_index
@@ -19,6 +20,10 @@ WEIGHTS = {
     "color": 0.3,
     "phash": 0.2
 }
+
+# Early termination toggle + chunking (kept internal; no signature change)
+_EARLY_TERMINATION = True
+_CHUNK_MULTIPLIER = 4  # submit work in chunks so we can prune between chunks
 
 
 def load_mapping(mapping_path):
@@ -79,7 +84,12 @@ def combined_similarity_search(
         if not db_entry:
             continue
         path, width, height = db_entry
-        candidates.append((path, clip_dist))
+        # Map Annoy angular distance to similarity (kept your existing mapping)
+        clip_sim = 1.0 - (clip_dist / 2.0)
+        candidates.append((path, clip_dist, clip_sim))
+
+    # Sort by CLIP similarity desc so our upper bound shrinks monotonically
+    candidates.sort(key=lambda x: x[2], reverse=True)
 
     # Parallel re-ranking (color + pHash) per candidate
     def _score_candidate(path: str, clip_dist: float):
@@ -89,7 +99,7 @@ def combined_similarity_search(
         candidate_img = preprocess_image(candidate_img)
 
         # CLIP similarity
-        clip_sim = 1.0 - (clip_dist / 2.0)  # angular [0,2] → similarity [1,0]
+        clip_sim_local = 1.0 - (clip_dist / 2.0)  # angular [0,2] → similarity [1,0]
 
         # Average color and pHash similarity across all query images
         color_sims = []
@@ -110,21 +120,49 @@ def combined_similarity_search(
 
         # Combined score
         combined = (
-            WEIGHTS["clip"] * clip_sim +
+            WEIGHTS["clip"] * clip_sim_local +
             WEIGHTS["color"] * avg_color_sim +
             WEIGHTS["phash"] * avg_phash_sim
         )
         return (path, combined)
 
-    scores = []
+    scores_heap = []  # min-heap of (combined, path)
     if candidates:
-        max_workers = min(multiprocessing.cpu_count(), len(candidates))
-        with ThreadPoolExecutor(max_workers=max_workers or 1) as ex:
-            futs = [ex.submit(_score_candidate, path, dist) for path, dist in candidates]
-            for fut in as_completed(futs):
-                res = fut.result()
-                if res:
-                    scores.append(res)
+        max_workers = min(multiprocessing.cpu_count(), len(candidates)) or 1
+        chunk_size = max_workers * _CHUNK_MULTIPLIER
 
-    scores.sort(key=lambda x: x[1], reverse=True)
-    return scores[:top_k_result]
+        i = 0
+        while i < len(candidates):
+            chunk = candidates[i:i + chunk_size]
+
+            with ThreadPoolExecutor(max_workers=max_workers) as ex:
+                futs = [ex.submit(_score_candidate, path, dist) for (path, dist, _sim) in chunk]
+                for fut in as_completed(futs):
+                    res = fut.result()
+                    if not res:
+                        continue
+                    path, combined = res
+                    if len(scores_heap) < top_k_result:
+                        heappush(scores_heap, (combined, path))
+                    else:
+                        heappushpop(scores_heap, (combined, path))
+
+            i += len(chunk)
+
+            # Early termination check (only if we already filled top-k)
+            if _EARLY_TERMINATION and len(scores_heap) >= top_k_result and i < len(candidates):
+                # Upper bound for any remaining candidate:
+                # assume color=1 and phash=1 (best possible), with next candidate's clip_sim.
+                next_clip_sim = candidates[i][2]
+                upper_bound = (
+                    WEIGHTS["clip"] * next_clip_sim +
+                    WEIGHTS["color"] * 1.0 +
+                    WEIGHTS["phash"] * 1.0
+                )
+                worst_in_topk = scores_heap[0][0]  # min in heap
+                if upper_bound <= worst_in_topk:
+                    break
+
+    # Convert heap to sorted list desc
+    top = nlargest(top_k_result, scores_heap)
+    return [(path, combined) for (combined, path) in top]