## Why

Index builds and per-query re-ranking were CPU-heavy and I/O bound. This PR:

* Avoids repeated CLIP model loads,
* Embeds images in batches for index builds,
* Re-ranks CLIP candidates in parallel,
* Stops work early when remaining candidates can’t beat current top-K.

Together these changes cut build and query times substantially (details below).

---

## What changed

### CLIP model & embeddings

* `similarity/similarity_embedding.py`

  * Added **`get_clip_model()`** (singleton cache) and **`compute_clip_embeddings_batch()`** (batched).
  * Kept existing `compute_clip_embedding` API unchanged. (Batch path uses the cached model under the hood.)

### Offline index build

* `pipeline/build_embedding_index.py`

  * Now **streams images** and **embeds in batches** (`BATCH_SIZE`), then writes into Annoy.
  * Minimal change; same CLI & outputs (`clip_index.ann`, `index_to_id.json`).

### Query-time re-ranking

* `pipeline/search_pipeline.py`

  * New **parallel re-ranking** of candidates (color + pHash) via `ThreadPoolExecutor`.
  * Added **early termination** using a fixed-size top-K heap and a CLIP-based upper bound. (Big wins when `k_clip` is large; no API changes.)

### Benchmarks & tests

* `image_recommender/tools/bench_clip_cache.py` – measures CLIP load vs. cached, embed timings.
* `image_recommender/tools/bench_clip_batch.py` – compares per-image vs. batched embeds.
* `tests/test_clip_model_cache.py` – validates singleton & embedding equivalence.
* `tests/test_clip_batch.py` – shape, dtype, norm=1, and batch≈single parity.
* `pytest.ini` – silence upstream `pkg_resources` deprecation warnings in tests.

---

## Results (on CPU, your machine)

From local runs shared in this branch:

* **CLIP model load**: `clip.load()` \~**1.67s** each → **\~0ms** after the first call using cache (mean amortized \~80ms over 20 runs).
* **Embed forward pass**: \~**45–50ms** (unchanged; same math).
* **End-to-end naive vs. cached**: \~**35×** faster when avoiding model reloads per call.
* **Batched index build**: expect **2–5×** speedup on CPU (more on GPU), depending on `BATCH_SIZE`.
* **Parallel re-rank**: with `k_clip ≥ 100`, **\~2–3×** faster; with **early termination**, **30–60%** less work when many candidates remain unlikely to enter top-K.

*(Exact numbers vary by disk, CPU, and image sizes.)*

---

## How to test

### Unit tests

```bash
pytest -q
```

### Cache benchmark

```bash
python image_recommender/tools/bench_clip_cache.py \
  --image image_recommender/data/sample.jpg --repeats 20
```

### Batch benchmark

```bash
python image_recommender/tools/bench_clip_batch.py
```

### Build index (batched)

```bash
python -m image_recommender.pipeline.build_embedding_index \
  --dataset /path/to/images \
  --index-out image_recommender/data/out/clip_index.ann \
  --mapping-out image_recommender/data/out/index_to_id.json \
  --batch-size 64 --n-trees 32
```

### Run search (parallel + early stop)

Use a larger `k_clip` (e.g., 200–1000) to see the gains; `top_k_result` stays small (e.g., 5–10).

---

## Config knobs

* **Batching**: `BATCH_SIZE` (index build).
* **Re-rank weights**: `WEIGHTS = {"clip": 0.5, "color": 0.3, "phash": 0.2}`.
* **Workers**: auto to `cpu_count()`, can be overridden.
* **Early termination**: `_EARLY_TERMINATION = True`, `_CHUNK_MULTIPLIER = 4` (internal constants).
* **Annoy build**: `n_trees` (higher = better recall, slower build).

---

## Backwards compatibility

* Public functions and CLIs remain the same.
* Mapping format for index build remains unchanged.
* Database helper is additive and unused by default.

---

## Risks & mitigations

* **Disk I/O saturation** from parallel image loads → bounded thread pool; chunked submissions.
* **Minor score variance** from early termination: only stops when an upper bound can’t beat current top-K.
* **Memory**: batch size configurable; only a batch of images resident during build.

---